ml_integration/django_ml/requirement.txt
django
djangorestframework
langchain
langchain_huggingface
chromadb
sentence-transformers
pandas
numpy
python-dotenv
